This commit aims to move the ComparableExtensions from the core project into the test project's helpers. Adjusting the unit tests to import the appropriate namespace.
